move wind to avatar.js
