+++ scenes/block.scn
@@ -116,15 +116,20 @@
     },
     {
       "position": [ 0, 5, -7 ],
-      "quaternion": [ 0, 0, 0, 1 ],
-      "start_url": "/avatars/drake_hacker_v3_vian.vrm",
+      "quaternion": [
+        0,
+        0,
+        0,
+        1
+      ],
+      "start_url": "/avatars/scillia_drophunter_v15_vian.vrm",
       "components": [
         {
           "key": "npc",
           "value": {
-            "name": "Drake",
-            "voice": "Shining Armor",
-            "bio": "His nickname is DRK. 15/M hacker. Loves guns. Likes plotting new hacks. He has the best equipment and is always ready for a fight."
+            "name": "Scillia",
+            "voice": "Sweetie Belle",
+            "bio": "Her nickname is Scilly or SLY. 13/F drop hunter. She is an adventurer, swordfighter and fan of potions. She is exceptionally skilled and can go Super Saiyan."
           }
         }
       ]
