+++ world.js
@@ -25,6 +25,7 @@ const localEuler = new THREE.Euler();
 
 // world
 export const world = {};
+world.winds = [];
 
 const appManager = new AppManager({
   appsMap: null,
